requirements.txt
flask
streamlit
requests
pdfplumber
pytesseract
Pillow
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
scikit-learn
python-dotenv
uuid
